pyproject.toml
[tool.poetry]
name = "deep-organize"
version = "0.1.0"
description = ""
authors = ["jloveric <john.loverich@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.2"
lion-pytorch = "^0.1.2"
torchmetrics = "^1.2.1"
lightning = "^2.1.3"
hydra-core = "^1.3.2"
matplotlib = "^3.8.2"
pillow = "^10.2.0"
torchvision = "^0.16.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
tensorboard = "^2.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
